bpmn_parser renamed to bpmn_json_parser
